Move mockpacakge tests to closure integration tests
