+++ changelog/pending/20240325--sdk-nodejs--move-mockpackage-tests-to-closure-integration-tests.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: chore
+  scope: sdk/nodejs
+  description: Move mockpackage tests to closure integration tests
