+++ sdk/nodejs/.eslintrc.js
@@ -16,7 +16,6 @@ module.exports = {
         "cmd/pulumi-language-nodejs",
         "tests/automation/data/tcfg/*.ts",
         "tests/sxs_ts_test/*.ts",
-        "tests/mockpackage/**/*.ts",
         "tests/runtime/testdata/closure-tests/**/*.ts",
         "tests/runtime/testdata/closure-tests/**/*.js",
     ],
