+++ sdk/nodejs/Makefile
@@ -35,8 +35,6 @@ build_package:: ensure
 	mkdir -p bin/tests/automation/data/
 	cp -R tests/automation/data/. bin/tests/automation/data/
 	cp README.md ../../LICENSE ./dist/* bin/
-# safely remove mockpackage dependency see [pulumi/pulumi#9026]
-	cp package.json yarn.lock bin/ && cd bin && yarn remove mockpackage && rm yarn.lock
 	node ../../scripts/reversion.js bin/package.json ${VERSION}
 	node ../../scripts/reversion.js bin/version.js ${VERSION}
 	mkdir -p bin/proto
