+++ sdk/nodejs/tests/runtime/testdata/closure-tests/mockpackage-src/lib/index.js
