+++ sdk/nodejs/tests/runtime/testdata/closure-tests/mockpackage-src/package.json
