--- sdk/nodejs/tests/runtime/testdata/closure-tests/mockpackage/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "mockpackage",
-  "version": "1.0.0",
-  "description": "",
-  "author": "",
-  "license": "ISC",
-  "exports": {
-    ".": {
-      "require": "./lib/index.js",
-      "import": "./lib/index.js"
-    }
-  },
-
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts"
-}
